fix: Outline font family was hardcoded
The font family for the outline component title was hard coded as
`Inter` but the base font family shipped with the docs doesn't include
`Inter` instead including `Inter var` or others.

This could be changed to be `font-family: var(--rp-font-family-base);`
but that seems redundant for this component where the only other places
that overwrite the font family are for code blocks.